/@/store/domain/stampHistoryの分離
